[D4CRIS-852] manage authority value (remove stored_value from authority place in solr fields)
